유저, 운동 기능 구현 (#2)
- Spring Security 적용
- Auth 도메인 구현
- Exercise 도메인 구현
- Preference 도메인 Auth와 Exercise 연동

+++ src/main/java/com/healthmate/healthmate/domain/preference/service/PreferenceService.java
@@ -10,5 +10,5 @@ public interface PreferenceService {
 	void updatePreference(Long id, UpdatePreferenceRequestDto req);
 	void deletePreference(Long id);
 	PreferenceResponseDto getPreference(Long id);
-	List<PreferenceResponseDto> getPreferencesByExercise(Integer exerciseId);
+	List<PreferenceResponseDto> getPreferencesByExercise(Long exerciseId);
 }
